requirements.txt
# Runtime dependencies
fastapi>=0.100
uvicorn[standard]>=0.22
streamlit>=1.30
pydantic>=2.0
requests>=2.31
python-dotenv>=1.0
alembic>=1.11
matplotlib>=3.7
plotly>=5.0.0
rq>=1.10
redis>=5.0
python-multipart>=0.0.6
email-validator>=1.3
typing-extensions>=4.7
pandas>=2.0
numpy>=1.23
scipy>=1.11
scikit-learn>=1.3
sqlalchemy
psycopg2-binary


# Development dependencies
pytest>=7.4
pytest-cov>=4.1
flake8>=6.0
black>=23.7
isort>=5.12
mypy>=1.4
